requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.2
pillow>=8.0.0
opencv-python>=4.5.1
albumentations>=0.5.2
tqdm>=4.50.0
tensorboard>=2.4.0
PyYAML>=5.4.1
scikit-learn>=0.24.1
matplotlib>=3.3.3 